Reparent nodes into detached DocumentFragment when clearing NodePart
Imagine this template structure:

```js
render(html`${
  // Outer NodePart
  ${
    html`${
      // Nested NodePart here
      directive()
    }`
  }
}`, container)
```

When the `NodePart` renders a `TemplateResult`, we get a structure somewhat like:

```html
<!--NodePart's Start-->
<!--    nested NodePart's Start-->
<!--    nested NodePart's End-->
<!--NodePart's End-->
```

This is because we take the `TemplateResult`'s template fragment (which has the nested `NodePart`'s start/end nodes as direct children), and inject it directly between the outer `NodePart`'s start/end nodes.

When we change the value that's rendered into that part, we would clear out everything between the `NodePart`'s. That leaves the `TemplateResult`'s nested `NodePart`'s start/end nodes completely detached (no `parentNode`). When the nested `NodePart` tries to update (via an async directive), we would try to mutate the nested `nodePart.startNode.parentNode` (which is `null`)!

This fixes it in a hacky way be reparenting those cleared nodes into a `DocumentFragment`. The nested `NodePart` will have now be attached to the `DocumentFragment`, meaning it can update without issue.

Fixes #293